swap

+++ .shell/prepare.sh
@@ -121,15 +121,22 @@ start_prepare()
     SWAP="/root/swap"
     if grep -q "use_swap = 2" /opt/config/mod_data/variables.cfg
         then
-        FREE_SPACE=$(df /media 2>/dev/null| tail -1 | tr -s ' ' | cut -d' ' -f4)
-        MIN_SPACE=$((128*1024))
+            for i in `seq 1 6`; do mount |grep /media && break; echo $i; sleep 10; done;
 
-        if [ "$FREE_SPACE" != "" ] && [ "$FREE_SPACE" -ge "$MIN_SPACE" ]
-            then
-                SWAP="/media/swap"
-                if ! [ -f $SWAP ]; then dd if=/dev/zero of=$SWAP bs=1024 count=131072; mkswap $SWAP; fi;
-                swapon $SWAP
-        fi
+            if mount |grep /media
+                then
+                    FREE_SPACE=$(df /media 2>/dev/null|grep -v /dev/root|grep -v Filesystem| tail -1 | tr -s ' ' | cut -d' ' -f4)
+                    MIN_SPACE=$((128*1024))
+                    mount
+                    df /media
+
+                    if [ "$FREE_SPACE" != "" ] && [ "$FREE_SPACE" -ge "$MIN_SPACE" ]
+                        then
+                            SWAP="/media/swap"
+                            if ! [ -f $SWAP ]; then dd if=/dev/zero of=$SWAP bs=1024 count=131072; mkswap $SWAP; fi;
+                            swapon $SWAP
+                    fi
+            fi
     fi
 
     chroot $MOD /opt/config/mod/.shell/root/start.sh "$SWAP" &
